split containers and views
