+++ app/assets/javascripts/components/settings/containers/add_admin_form_container.jsx
@@ -1,0 +1,13 @@
+import { connect } from 'react-redux';
+import { upgradeAdmin } from '../../../actions/settings_actions';
+import AddAdminForm from '../views/add_admin_form';
+
+const mapStateToProps = state => ({
+  submittingNewAdmin: state.settings.submittingNewAdmin,
+});
+
+const mapDispatchToProps = {
+  upgradeAdmin,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(AddAdminForm);
