+++ app/assets/javascripts/components/settings/settings_handler.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { fetchAdminUsers } from '../../actions/settings_actions';
 import { connect } from 'react-redux';
 import AdminUserList from './admin_users_list.jsx';
-import AddAdminButton from './add_admin_button.jsx';
+import AddAdminButton from './views/add_admin_button.jsx';
 import Notifications from '../common/notifications';
 
 class SettingsHandler extends React.Component {
